// frontend/magic-web/src/opensource/pages/superMagic/components/Detail/components/FilesViewer/types.ts
export interface FileItem {
	file_id: string
	file_name: string
	display_filename?: string
	filename?: string
	file_extension?: string
	file_url?: string
	relative_file_path?: string
	url?: string
	is_directory?: boolean
	children?: FileItem[]
	content?: string
	updated_at?: string
	metadata?: {
		type?: "slide" | "design" | "dashboard" | "audio" | "video" | string
		name?: string
	}
	file_size?: number
	parent_id?: string | number
	source?: AttachmentSource
	project_id?: string
}